adjusted ipg time

+++ linux/avb/avb_1722avtp.c
@@ -484,7 +484,7 @@ int init_1722_driver( ieee1722_avtp_driver_state_t *ieee1722mc, const char* name
 
     ieee1722mc->sample_rate = sample_rate;
     ieee1722mc->period_size = period_size;
-    ieee1722mc->period_usecs = (uint64_t) (period_size / sample_rate * 1000000);
+    ieee1722mc->period_usecs = (uint64_t) ((float)period_size / (float)sample_rate * 1000000);
 
     fprintf(filepointer,"sample_rate: %d, period size: %d, period usec: %lld\n",ieee1722mc->sample_rate, ieee1722mc->period_size, ieee1722mc->period_usecs);fflush(filepointer);
